pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "my-larapy-project"
version = "0.1.0"
description = "A Larapy application"
requires-python = ">=3.11"
license = {text = "MIT"}

dependencies = [
    "larapy-framework>=0.9.0,<1.0.0",
    "uvicorn[standard]>=0.24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "black>=24.0.0",
    "ruff>=0.1.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]

[tool.black]
line-length = 100
target-version = ["py311", "py312", "py313"]

[tool.ruff]
line-length = 100
target-version = "py311"
